+++ docs/changelog.rst
@@ -8,6 +8,7 @@ v1.0.0 - *not yet released*
     - Drop support for Python 3.9, begin testing on Python 3.14
     - **REMOVED** :func:`~igbpyutils.dev.generate_coveragerc`
         - Use https://pypi.org/project/coverage-simple-excludes/ instead.
+    - ``igbpyutils.dev.check_script_vs_lib`` now actually raises a deprecation warning.
 
 v0.8.0 - Fri Oct 11 2024
     - **REMOVED** ``gen-coveragerc`` and **deprecated** :func:`~igbpyutils.dev.generate_coveragerc`